Moved DBStatementCounter and PerfTimer to utils
This is for more modular code and a structure that actually makes sense
